requirements.txt

bert-score
blobfile
datasets
huggingface-hub==0.4.0
mpi4py
nltk
numpy              
pandas 
protobuf 
rouge-score  
sacrebleu   
sacremoses   
scikit-learn  
scipy 
spacy     
tokenizers    
torchmetrics  
tqdm     
transformers==4.18.0